==> notebooks/2_gct_analyze/05_gct_export.ipynb <==
%load_ext autoreload
%autoreload 2
import sys
from ridge_utils.DataSequence import DataSequence
import pandas as pd
import os
from os.path import join
from collections import defaultdict
import numpy as np
import joblib
from tqdm import tqdm
from sasc.config import FMRI_DIR, STORIES_DIR, RESULTS_DIR
from neuro.config import brain_drive_resps_dir
import ridge_utils.DataSequence
import DataSequence

import numpy as np
import itertools as itools
from ridge_utils.interpdata import sincinterp2D, gabor_xfm2D, lanczosinterp2D

class DataSequence(object):
    """DataSequence class provides a nice interface for handling data that is both continuous
    and discretely chunked. For example, semantic projections of speech stimuli must be
    considered both at the level of single words (which are continuous throughout the stimulus)
    and at the level of TRs (which contain discrete chunks of words).
    """
    def __init__(self, data, split_inds, data_times=None, tr_times=None):
        """Initializes the DataSequence with the given [data] object (which can be any iterable)
        and a collection of [split_inds], which should be the indices where the data is split into
        separate TR chunks.
        """
        self.data = data
        self.split_inds = split_inds
        self.data_times = data_times
        self.tr_times = tr_times

    def mapdata(self, fun):
        """Creates a new DataSequence where each element of [data] is produced by mapping the
        function [fun] onto this DataSequence's [data].

        The [split_inds] are preserved exactly.
        """
        return DataSequence(self, list(map(fun, self.data)), self.split_inds)

    def chunks(self):
        """Splits the stored [data] into the discrete chunks and returns them.
        """
        return np.split(self.data, self.split_inds)

    def data_to_chunk_ind(self, dataind):
        """Returns the index of the chunk containing the data with the given index.
        """
        zc = np.zeros((len(self.data),))
        zc[dataind] = 1.0
        ch = np.array([ch.sum() for ch in np.split(zc, self.split_inds)])
        return np.nonzero(ch)[0][0]

    def chunk_to_data_ind(self, chunkind):
        """Returns the indexes of the data contained in the chunk with the given index.
        """
        return list(np.split(np.arange(len(self.data)), self.split_inds)[chunkind])

    def chunkmeans(self):
        """Splits the stored [data] into the discrete chunks, then takes the mean of each chunk
        (this is assuming that [data] is a numpy array) and returns the resulting matrix with
        one row per chunk.
        """
        dsize = self.data.shape[1]
        outmat = np.zeros((len(self.split_inds)+1, dsize))
        for ci, c in enumerate(self.chunks()):
            if len(c):
                outmat[ci] = np.vstack(c).mean(0)

        return outmat

    def chunksums(self, interp="rect", **kwargs):
        """Splits the stored [data] into the discrete chunks, then takes the sum of each chunk
        (this is assuming that [data] is a numpy array) and returns the resulting matrix with
        one row per chunk.

        If [interp] is "sinc", the signal will be downsampled using a truncated sinc filter
        instead of a rectangular filter.

        if [interp] is "lanczos", the signal will be downsampled using a Lanczos filter.

        [kwargs] are passed to the interpolation function.
        """
        if interp=="sinc":
            ## downsample using sinc filter
            return sincinterp2D(self.data, self.data_times, self.tr_times, **kwargs)
        elif interp=="lanczos":
            ## downsample using Lanczos filter
            return lanczosinterp2D(self.data, self.data_times, self.tr_times, **kwargs)
        elif interp=="gabor":
            ## downsample using Gabor filter
            return np.abs(gabor_xfm2D(self.data.T, self.data_times, self.tr_times, **kwargs)).T
        else:
            dsize = self.data.shape[1]
            outmat = np.zeros((len(self.split_inds)+1, dsize))
            for ci, c in enumerate(self.chunks()):
                if len(c):
                    outmat[ci] = np.vstack(c).sum(0)
                    
            return outmat

    def copy(self):
        """Returns a copy of this DataSequence.
        """
        return DataSequence(list(self.data), self.split_inds.copy(), self.data_times, self.tr_times)
    
    @classmethod
    def from_grid(cls, grid_transcript, trfile):
        """Creates a new DataSequence from a [grid_transript] and a [trfile].
        grid_transcript should be the product of the 'make_simple_transcript' method of TextGrid.
        """
        data_entries = list(zip(*grid_transcript))[2]
        if isinstance(data_entries[0], str):
            data = list(map(str.lower, list(zip(*grid_transcript))[2]))
        else:
            data = data_entries
        word_starts = np.array(list(map(float, list(zip(*grid_transcript))[0])))
        word_ends = np.array(list(map(float, list(zip(*grid_transcript))[1])))
        word_avgtimes = (word_starts + word_ends)/2.0
        
        tr = trfile.avgtr
        trtimes = trfile.get_reltriggertimes()
        
        split_inds = [(word_starts<(t+tr)).sum() for t in trtimes][:-1]
        return cls(data, split_inds, word_avgtimes, trtimes+tr/2.0)

    @classmethod
    def from_chunks(cls, chunks):
        """The inverse operation of DataSequence.chunks(), this function concatenates
        the [chunks] and infers split_inds.
        """
        lens = list(map(len, chunks))
        split_inds = np.cumsum(lens)[:-1]
        #data = reduce(list.__add__, map(list, chunks)) ## 2.26s for 10k 6-w chunks
        data = list(itools.chain(*map(list, chunks))) ## 19.6ms for 10k 6-w chunks
        return cls(data, split_inds)


wordseqs_dir = join(RESULTS_DIR, 'processed', 'wordseqs')

wordseqs = [
    joblib.load(join(wordseqs_dir, f)) for f in os.listdir(wordseqs_dir)
]
# concatenate list of dicts
wordseqs_dict = {}
for d in wordseqs:
    for k, v in d.items():
        wordseqs_dict[k] = ridge_utils.DataSequence.DataSequence(
            data=v.data, split_inds=v.split_inds, data_times=v.data_times, tr_times=v.tr_times)

wordseqs_dict

RESPS_DICT = {
    '20230504': {
        "default/uts02_pilot_gpt4_mar28___ver=v4_noun___seed=1": "GenStory1_resps.npy",
        "default/uts02_pilot_gpt4_mar28___ver=v4_noun___seed=3": "GenStory2_resps.npy",
        "default/uts02_pilot_gpt4_mar28___ver=v4_noun___seed=4": "GenStory3_resps.npy",
        "default/uts02_pilot_gpt4_mar28___ver=v5_noun___seed=1": "GenStory4_resps.npy",
        "default/uts02_pilot_gpt4_mar28___ver=v5_noun___seed=2": "GenStory5_resps.npy",
        "default/uts02_pilot_gpt4_mar28___ver=v5_noun___seed=4": "GenStory6_resps.npy",
    },
    '20230702': {
        "interactions/uts02___jun14___seed=1": "GenStory7_resps.npy",
        "interactions/uts02___jun14___seed=4": "GenStory8_resps.npy",
        "polysemantic/uts02___jun14___seed=6": "GenStory9_resps.npy",
        "polysemantic/uts02___jun14___seed=1": "GenStory10_resps.npy",
    },
    '20231106': {
        'default/uts03___jun14___seed=5': 'GenStory12_resps.npy',
        'default/uts03___jun14___seed=1': 'GenStory13_resps.npy',
        'interactions/uts03___jun14___seed=5': 'GenStory14_resps.npy',
        'interactions/uts03___jun14___seed=6': 'GenStory15_resps.npy',
        'polysemantic/uts03___jun14___seed=3': 'GenStory16_resps.npy',
        'polysemantic/uts03___jun14___seed=7': 'GenStory17_resps.npy',
    },
    '20240509': {
        'default/uts01___may9___seed=5_top1': 'deeptune-story19.npy',
        'default/uts01___may9___seed=2_top2': 'deeptune-story20.npy',
        'interactions/uts01___may9___seed=3_top1': 'deeptune-story21.npy',
        'interactions/uts01___may9___seed=6_top2': 'deeptune-story22.npy',
    },

    '20240604': {
        'roi/uts02___roi_may31___seed=5_best1': 'GenStory23.npy',
        'roi/uts02___roi_may31___seed=2_best2': 'GenStory24.npy',
        'roi/uts02___roi_may31___seed=7_best3': 'GenStory25.npy',
        'roi/uts02___roi_may31___seed=6_best4': 'GenStory26.npy',

        'qa/uts02___qa_may31___seed=1': 'GenStory27_resps.npy',
        'qa/uts02___qa_may31___seed=2': 'GenStory28_resps.npy',
        'qa/uts02___qa_may31___seed=3': 'GenStory29_resps.npy',
    },
}


def build_wordseq(timings):
    TR_TIME = 2
    words = timings['word'].values
    word_lengths = timings['timing'].values
    end_times = timings['time_running'].values

    # Compute the average times of the words (halfway through each word)
    word_avgtimes = end_times - (word_lengths / 2.0)

    # Compute the TR times starting from -9
    tr_times = np.arange(-9, word_avgtimes[-1] + TR_TIME, TR_TIME)

    # Compute the split indices
    split_inds = [(word_avgtimes < (t + TR_TIME)).sum() for t in tr_times][:-1]

    ds = DataSequence(words, split_inds, word_avgtimes, tr_times)
    return ds

def get_num_from_string(s):
    return int(''.join(filter(str.isdigit, s)))


dset = defaultdict(list)
for session in tqdm(RESPS_DICT.keys()):
    for k, v in RESPS_DICT[session].items():
        dset['session'].append(session)
        dset['story_name'].append(k)
        dset['resp_file'].append(v.replace('_resps', ''))
        # all the time comes from loading the resp here
        resp = np.load(join(brain_drive_resps_dir,
                       session, dset['resp_file'][-1]))
        dset['trs_resp'].append(resp.shape[0])

        if session == '20240509':
            timings = pd.read_csv(
                join(STORIES_DIR, k, 'timings_processed_slowed.csv'))
            # display(timings.head())
        else:
            timings = pd.read_csv(
                join(STORIES_DIR, k, 'timings_processed.csv'))

        # get number from string
        v_string = get_num_from_string(v)
        wordseq = wordseqs_dict[f'GenStory{v_string}']

        # wordseq = build_wordseq(timings)
        if len(wordseq.tr_times) - dset['trs_resp'][-1] == 16:
            # trim off last tr from wordseq
            wordseq.tr_times = wordseq.tr_times[:-1]
        dset['trs_timings'].append(int(max(timings['time_running']) // 2))
        dset['trs_wordseq'].append(len(wordseq.tr_times))
        dset['trs_diff'].append(dset['trs_wordseq'][-1] - dset['trs_resp'][-1])
        dset['subject'].append(k.split('/')[1].split('_')[0].upper())
        dset['wordseq'].append(wordseq)

df = pd.DataFrame(dset)

df['story_id'] = 'GenStory' + df['resp_file'].str.replace(
    '.npy', '').str.replace('GenStory', '').str.replace('deeptune-story', '')
df = df.set_index('story_id')

df

joblib.dump(df, join(brain_drive_resps_dir, 'metadata.pkl'))

df[df.subject == 'UTS03'].story_id.values

df.loc[['GenStory5', 'GenStory7']]['wordseq'].to_dict()
